+++ ui/packages/components/src/Table/index.ts
@@ -1,3 +1,4 @@
 export { Table } from './Table';
 export { TableFilter } from './Filter';
 export { IDCell, TextCell, TimeCell, StatusCell, PillCell, NumberCell } from './Cell';
+export { TableBlankState } from './TableBlankState';
